fix: use system context for querying workspaces when deleting users (#19211) (#19227)
Backport of #19211 to our ESR, v2.24, since this is a kind of bad
(albeit rare) bug and very easy to fix.

### Original PR

Closes #19209.

In `templates.go`, we do this to make sure we count ALL workspaces for a
template before we try and delete that template:
https://github.com/coder/coder/blob/dc598856e3be0926573dbbe2ec680e95a139093a/coderd/templates.go#L81-L99

However, we weren't doing the same when attempting to delete users,
leading to the linked issue. We can solve the issue the same way as we
do for templates.