fix: increase timeout for watch workspace agent devcontainers test
Fixes https://github.com/coder/internal/issues/907

When running in a CI environment, the process of setting up the agent +
coderd + creating the websocket connection can sometimes take up to 10
seconds, by which time we timeout. To resolve this we up the timeout
from `WaitShort` to `WaitLong`.

fix: run the formatter

fix: WaitSuperLong for watch containers PayloadTooLarge test

The test can take around 10s when it is the only one running, so in a
constrained environment like CI it makes sense that it still hits the 25
second timeout. For now we up the limit to 60 seconds until the test is
rewritten to greatly reduce the time taken.