chore: update docs to fix ordering of OS options
On https://coder.com/docs/install/uninstall at present we order the top OS listing as "Linux | macOS | Windows", while in the `Coder settings, cache, and the optional built-in PostgreSQL database` paragraph towards the bottom of the page we change to using "macOS | Linux | Windows" for some reason. This PR moves Linux to be listed first instead of macOS in the bottom paragraph to match the ordering of the top section.