fix(coderd): ignore sub agents when converting a task to workspace (#19624)
Addresses comment raised on previous PR
https://github.com/coder/coder/pull/19619#discussion_r2307943410

We know we can skip sub agents when searching for which agent is related
to the task, as this is not an explicitly supported feature at the
moment. When we come to properly setting up a Task -> Agent relationship
this limitation will be dropped.