fix(coderd): allow fetching a newly created task
Due to how we currently label a workspace as a task, there is a delay
between when a task workspace is created and when it is labelled as
a task.

This PR introduces fallback check for when a workspace does _not_ have
`HasAITask` set. This fallback check tests to see if the special "AI
Prompt" parameter is present in the workspace's build parameters.